Separate inventory selectors into their own file
