+++ src/app/collections/Collections.tsx
@@ -14,7 +14,7 @@ import { connect } from 'react-redux';
 import { InventoryBuckets } from '../inventory/inventory-buckets';
 import { RootState } from '../store/reducers';
 import { createSelector } from 'reselect';
-import { storesSelector, profileResponseSelector } from '../inventory/reducer';
+import { storesSelector, profileResponseSelector } from '../inventory/selectors';
 import { refresh$ } from '../shell/refresh';
 import PresentationNodeRoot from './PresentationNodeRoot';
 import Mods from './Mods';
