+++ src/app/destiny1/loadout-builder/D1LoadoutBuilder.tsx
@@ -15,7 +15,7 @@ import { connect } from 'react-redux';
 import { DestinyAccount } from '../../accounts/destiny-account';
 import { D1Store } from '../../inventory/store-types';
 import { RootState } from '../../store/reducers';
-import { storesSelector } from '../../inventory/reducer';
+import { storesSelector } from '../../inventory/selectors';
 import { currentAccountSelector } from '../../accounts/reducer';
 import { D1StoresService } from '../../inventory/d1-stores';
 import { Loading } from '../../dim-ui/Loading';
