+++ src/app/farming/reducer.ts
@@ -3,7 +3,7 @@ import * as actions from './actions';
 import { ActionType, getType } from 'typesafe-actions';
 import _ from 'lodash';
 import { createSelector } from 'reselect';
-import { storesSelector } from '../inventory/reducer';
+import { storesSelector } from '../inventory/selectors';
 import { RootState } from '../store/reducers';
 
 export const farmingStoreSelector = () =>
