+++ src/app/organizer/ItemTable.tsx
@@ -32,7 +32,7 @@ import { bulkTagItems } from 'app/inventory/tag-items';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
 import { RootState, ThunkDispatchProp } from 'app/store/reducers';
-import { storesSelector, itemInfosSelector } from 'app/inventory/reducer';
+import { storesSelector, itemInfosSelector } from 'app/inventory/selectors';
 import { searchFilterSelector } from 'app/search/search-filters';
 import { inventoryWishListsSelector } from 'app/wishlists/reducer';
 import { toggleSearchQueryComponent } from 'app/shell/actions';
